Port tests for Nullable to ava
