--- __tests__/Nullable/or.test.js
@@ -1,43 +1,0 @@
-'use strict';
-
-const assert = require('assert');
-
-const { orForNullable } = require('../../__dist/cjs/Nullable/or');
-const { nonNullableValue } = require('../utils');
-
-describe('Nullable::or', () => {
-    const LEFT = Symbol('a');
-    const RIGHT = Symbol('b');
-
-    const nonNullableValueList = [];
-    for (const value of nonNullableValue) {
-        const LEFT = Symbol('a');
-        const RIGHT = Symbol('b');
-        nonNullableValueList.push([value, RIGHT, value]);
-        nonNullableValueList.push([LEFT, value, LEFT]);
-        nonNullableValueList.push([value, value, value]);
-        nonNullableValueList.push([null, value, value]);
-        nonNullableValueList.push([value, null, value]);
-        nonNullableValueList.push([undefined, value, undefined]);
-        nonNullableValueList.push([value, undefined, value]);
-    }
-
-    const list = [
-        ...nonNullableValueList,
-        [null, RIGHT, RIGHT],
-        [undefined, RIGHT, undefined],
-        [LEFT, RIGHT, LEFT],
-        [LEFT, null, LEFT],
-        [LEFT, undefined, LEFT],
-        [null, null, null],
-        [null, undefined, undefined],
-        [undefined, undefined, undefined],
-        [undefined, null, undefined],
-    ];
-
-    for (const [a, b, expected] of list) {
-        it(`a=${String(a)}, b=${String(b)}`, () => {
-            assert.strictEqual(orForNullable(a, b), expected);
-        });
-    }
-});
