+++ ava_test/Nullable/tap.test.js
@@ -1,0 +1,39 @@
+import test from 'ava';
+
+const { tapNullable } = require('../../__dist/cjs/Nullable/tap');
+const { nonNullableValue } = require('../utils');
+
+for (const input of nonNullableValue) {
+    test('pass the value: ' + String(input), (t) => {
+        t.plan(3);
+
+        const result = tapNullable(input, (v) => {
+            t.pass();
+            t.is(v, input, 'the arg is input');
+        });
+
+        t.is(result, input);
+    });
+}
+
+test('pass null', (t) => {
+    t.plan(1);
+
+    const result = tapNullable(null, (_v) => {
+        t.fail();
+    });
+    t.is(result, null);
+});
+
+test('pass undefined', (t) => {
+    t.plan(3);
+
+    const INPUT = undefined;
+
+    const result = tapNullable(INPUT, (v) => {
+        t.pass();
+        t.is(v, INPUT, 'the arg is INPUT');
+    });
+
+    t.is(result, INPUT);
+});
